+++ 692/polymorphism/Triangle.cpp
@@ -1,6 +1,6 @@
 #include "Triangle.h"
 
-Triangle::Triangle(double base, double height) : m_base(base), m_height(height)
+Triangle::Triangle(int base, int height) : m_base(base), m_height(height)
 {
 }
 
@@ -8,7 +8,7 @@ Triangle::~Triangle()
 {
 }
 
-double Triangle::area() 
+int Triangle::area() 
 {
   return (m_base * m_height / 2);
 }
